[fix] 패키지명 수정 [feat] xml 파일 생성, Mapper interface 생성
