+++ build.gradle
@@ -37,6 +37,10 @@ dependencies {
 	implementation 'org.json:json:20240303'
 	implementation 'org.hibernate:hibernate-validator:6.2.0.Final'  // 최신 버전 확인
 
+	// MyBatis
+	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter:3.0.3'
+	implementation group: 'org.mybatis.dynamic-sql', name: 'mybatis-dynamic-sql', version: '1.5.2'
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 }
 
 tasks.named('test') {
