--- src/main/java/ggyuel/ggyuup/Member/service/MemberService.java
@@ -1,7 +1,0 @@
-package ggyuel.ggyuup.Member.service;
-
-import ggyuel.ggyuup.Member.dto.LoginDTO;
-
-public interface MemberService {
-    Boolean checkEwhain(LoginDTO request);
-}
