--- src/main/java/ggyuel/ggyuup/Organizations/domain/Organizations.java
@@ -1,13 +1,0 @@
-package ggyuel.ggyuup.Organizations.domain;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Organizations {
-    private int ranking;
-    private String groupName;
-    private int solvedNum;
-}
