--- src/main/java/ggyuel/ggyuup/Problems/domain/Problems.java
@@ -1,15 +1,0 @@
-package ggyuel.ggyuup.Problems.domain;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Problems {
-    private int problemId;
-    private String title;
-    private String link;
-    private int tier;
-    private int solvedNum;
-}
