--- src/main/java/ggyuel/ggyuup/TodayPs/domain/TodayPs.java
@@ -1,11 +1,0 @@
-package ggyuel.ggyuup.TodayPs.domain;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class TodayPs {
-    private int problemId;
-}
