+++ src/main/java/ggyuel/ggyuup/dataCrawling/repository/DataCrawlingMapper.java
@@ -1,0 +1,7 @@
+package ggyuel.ggyuup.dataCrawling.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface DataCrawlingMapper {
+}
