+++ src/main/java/ggyuel/ggyuup/member/dto/LoginDTO.java
@@ -1,0 +1,13 @@
+package ggyuel.ggyuup.member.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LoginDTO {
+
+    //  handle : 사용자의 백준 핸들
+    private String handle;
+
+}
