+++ src/main/java/ggyuel/ggyuup/member/service/MemberService.java
@@ -1,0 +1,7 @@
+package ggyuel.ggyuup.member.service;
+
+import ggyuel.ggyuup.member.dto.LoginDTO;
+
+public interface MemberService {
+    Boolean checkEwhain(LoginDTO request);
+}
