+++ src/main/java/ggyuel/ggyuup/problem/domain/Problems.java
@@ -1,0 +1,15 @@
+package ggyuel.ggyuup.problem.domain;
+
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Problems {
+    private int problemId;
+    private String title;
+    private String link;
+    private int tier;
+    private int solvedNum;
+}
