+++ src/main/java/ggyuel/ggyuup/problem/repository/ProblemMapper.java
@@ -1,0 +1,14 @@
+package ggyuel.ggyuup.problem.repository;
+
+import ggyuel.ggyuup.problem.dto.ProblemAlgoRespDTO;
+import ggyuel.ggyuup.problem.dto.ProblemTierRespDTO;
+import org.apache.ibatis.annotations.Mapper;
+import org.springframework.data.repository.query.Param;
+
+import java.util.List;
+
+@Mapper
+public interface ProblemMapper {
+    List<ProblemAlgoRespDTO> selectProblemsByAlgo(@Param("algoId") String algoId);
+    List<ProblemTierRespDTO> selectProblemsByTier(@Param("tier") int tier);
+}
