+++ src/main/java/ggyuel/ggyuup/todayPs/domain/TodayPs.java
@@ -1,0 +1,11 @@
+package ggyuel.ggyuup.todayPs.domain;
+
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class TodayPs {
+    private int problemId;
+}
