+++ src/main/resources/mappers/dataCrawling/DataCrawlingMapper.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="ggyuel.ggyuup.dataCrawling.repository.DataCrawlingMapper">
+
+</mapper>
