+++ src/main/resources/mappers/member/Member.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="ggyuel.ggyuup.member.repository.MemberMapper">
+    <select id="checkEwhain" resultType="int">
+        SELECT EXISTS (
+        SELECT 1
+        FROM students
+        WHERE handle = #{handle}
+        ) AS isExist
+    </select>
+</mapper>
